% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_abs_path.R
\name{file_abs_path}
\alias{file_abs_path}
\title{Absolute file path}
\usage{
file_abs_path(path, winslash = "\\\\", mustWork = NA)
}
\arguments{
\item{path}{characterl vector of paths}

\item{winslash}{character; separator to by used on windows}

\item{mustWork}{logical; if \code{TRUE} then an error is given if the result 
cannot be determined; if NA then a warning.}
}
\description{
Get the absolute file path; alias of \code{\link[base]{normalizePath}}.
}
\details{
See \code{\link[base]{normalizePath}}
}
