% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_basename.R
\name{file_basename}
\alias{file_basename}
\title{file_basename - get basename, optionally omitting extension}
\usage{
file_basename(path, extension = TRUE)
}
\arguments{
\item{path}{character; path of a resource}

\item{extension}{logical; whether to keep the file extension (default:TRUE) 

Fumctions as \code{\link[base]{basename}} except for the ability
to strip the extension from the path. The extension is the final
\code{.} character and anything that follows.}
}
\description{
Similar to \code{base::basename} but allows for the stripping 
of the extension, otherwise a complete drop-in replacements
}
\examples{
  file_basename( "foo/bar.baz" )
  file_basename( "foo/bar.baz", extension=FALSE )
  file_basename( "/foo/bar.baz.burp", extension=FALSE )
  file_basename( "/foo/bar.baz.", extension=FALSE )

}
