% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_extension.R
\name{file_extension}
\alias{file_extension}
\alias{file_ext}
\title{Get file extension}
\usage{
file_extension(x)

file_ext(x)
}
\arguments{
\item{x}{character vector of filenames}
}
\value{
character
}
\description{
Extract the extension from filename string
}
\details{
\itemize{
\item tk
}
}
\examples{
  
  file_extension("README.md")
  file_extension("README.txt")
  file_extension("TODO.text")
     
}
\seealso{
\code{stringr::str_extract}
}
