% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_glob.R
\name{file_glob}
\alias{file_glob}
\title{file globbing}
\usage{
file_glob(paths, dirmark = FALSE)
}
\arguments{
\item{paths}{character; vector of patterns for relative or absolute filepaths. 
Missing values are ignored.}

\item{dirmark}{logical; should matches to directories from patterns that do 
not already end in / have a slash appended? May not be supported on all 
platforms.}
}
\description{
Alias for \code{Sys.glob}
}
\details{
See \code{\link[base]{Sys.glob}}
}
