% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_mtime.R
\name{file_mtime}
\alias{file_mtime}
\title{modify mtime}
\usage{
file_mtime(path, value = NULL)
}
\arguments{
\item{path}{string; path to file}

\item{value}{mtime}

\item{...}{character vector of file paths}
}
\description{
modify file mtime
}
\details{
touch -t datetime -m path
}
\examples{

file_mtime("delete.me", now() ) 

"delete-me" \%>\% file_mtime( Sys.time() )

}
