% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_path.R
\name{file_path}
\alias{file_path}
\alias{\%/\%}
\title{file_path}
\usage{
file_path(...)

path \%/\% file
}
\arguments{
\item{...}{arguments passed \code{base::file.path}

\code{file_path`` and aliases}base::file.path``}

\item{path}{character}

\item{file}{character}
}
\description{
Alias for file.path, consistent with file.tools, plus tidy, intuitive syntax
for creating a file.path
}
\examples{
  file_path( "/path", "to", "my", "files" )
  
  "path" \%/\% "file"
  "path" \%/\% letters[1:3]
  
}
