% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_read.R
\name{file_read}
\alias{file_read}
\alias{read_file}
\title{file_read}
\usage{
file_read(x, ...)

read_file(...)
}
\arguments{
\item{x}{string; filename}

\item{...}{additional args

Reads in a file using internal \code{readxl::read_excel} or 
\code{file_read_encoded}.}
}
\description{
read text or excel file based on the file extension
}
\seealso{
\code{\link[readxl]{read_excel}} 
  \code{\link{file_read_encoded}}
}
