% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_read_encoded.R
\name{file_read_encoded}
\alias{file_read_encoded}
\alias{read_encoded}
\title{Read encoded files}
\usage{
file_read_encoded(file, encodings, ...)

read_encoded(...)
}
\arguments{
\item{file}{string;}

\item{encodings}{character; series of encodings to try.}

\item{...}{additional arguments

\code{file_read_encoded} is useful when you don't know the encoding of a file
and have to try a number of encodings.}
}
\description{
Try reading a file using a number of encodings
}
\examples{

\dontrun{
   "./my_encoded_file.txt" \%>\% file_read_encoded
}

}
