% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_wait.R
\name{file_wait}
\alias{file_wait}
\title{Wait on file modifiction}
\usage{
file_wait(file, time = 0.2)
}
\arguments{
\item{file}{path to file to be watched}

\item{time}{The time interval to suspend execution for, in seconds.

Simple, platform-independent file watcher. System is slept waiting for the 
modification time to change.}
}
\value{
NULL; used for side-effects only
}
\description{
Wait for a file to be modified
}
\seealso{
\code{\link[base]{Sys.sleep}} \cr
  \code{\link[base]{file.mtime}}
}
