% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_write_tsv.R
\name{file_write_tsv}
\alias{file_write_tsv}
\alias{write.tsv}
\title{file_write_tsv}
\usage{
file_write_tsv(x, file = "", ..., sep = "\\t", row.names = FALSE,
  quote = FALSE)

write.tsv(...)
}
\arguments{
\item{x}{object; he object to be written, preferably a matrix or data frame. 
If not, it is attempted to coerce x to a data frame.}

\item{file}{character string naming a file or a connection open for 
writing. "" indicates output to the console.}

\item{...}{additional arguments passed to \code{\link[utils]{write.table}}}

\item{sep}{string; Values within each row of x are separated by this string.}

\item{row.names}{logical or character, see \code{\link[utils]{write.table}}}

\item{quote}{logical or numeric, see \code{\link[utils]{write.table}}}
}
\description{
A wrapper around \code{\link[utils]{write.table}} with defaults suitable for
writing
}
\details{
\code{file_write_tsv} wraps \code{\link{write.table}} providing sensible 
  defaults for writing tab-separated value files.
}
\note{
shoule the default for file be a file named after x?
  file=paste0( substitute(x), ".tsv")
}
\seealso{
\code{\link[utils]{write.table}}
}
